mito-ai: rename file
