+++ mito-ai/mito_ai/completions/completion_handlers/chat_completion_handler.py
@@ -6,7 +6,7 @@ from typing import List, Union, AsyncGenerator, Callable
 from openai.types.chat import ChatCompletionMessageParam
 from mito_ai.completions.models import ChatMessageMetadata, MessageType, CompletionRequest, CompletionStreamChunk, CompletionReply
 from mito_ai.completions.prompt_builders.chat_prompt import create_chat_prompt
-from mito_ai.completions.providers import ProviderManager
+from mito_ai.provider_manager import ProviderManager
 from mito_ai.completions.message_history import GlobalMessageHistory
 from mito_ai.completions.completion_handlers.completion_handler import CompletionHandler
 from mito_ai.completions.completion_handlers.utils import append_chat_system_message, create_ai_optimized_message
